refactor(message-component): Rename file `message-type.enum` to `message-type.type`
The message type is a type and not an Enum. As such, the file was renamed to reflect its content properly.
