+++ src/app/shared/message/index.ts
@@ -1,4 +1,4 @@
 export { Message } from './message.model';
 export { MessageComponent } from './message.component';
 export { MessageService } from './message.service';
-export { MessageType } from './message-type.enum';
+export { MessageType } from './message-type.type';
