+++ src/app/shared/message/message.model.ts
@@ -1,4 +1,4 @@
-import { MessageType } from './message-type.enum';
+import { MessageType } from './message-type.type';
 export class Message {
   constructor(public type?: MessageType, public title?: String, public description?: String) {}
 }
